split failing test to ensure isolation
